Refactoring code to create separate Transaction component
